updates for services

+++ Backend/src/main/kotlin/corp/siendev/com/rickandmorty/service/CharacterService.kt
@@ -2,13 +2,14 @@ package corp.siendev.com.rickandmorty.service
 
 import corp.siendev.com.rickandmorty.entity.Character
 import corp.siendev.com.rickandmorty.repository.CharacterRepository
+import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
-import java.util.Optional
 
 @Service
 class CharacterService(private val characterRepository: CharacterRepository) {
-    fun getCharacterByName(characterName: String): Optional<Character> {
+    fun getCharacterByName(characterName: String): Character {
         return characterRepository.findCharacterByName(characterName)
+            .orElseThrow { EntityNotFoundException("Couldn't find character by name: $characterName") }
     }
 
     fun getAllCharactersBySpecies(species: String): List<Character> {
@@ -19,8 +20,9 @@ class CharacterService(private val characterRepository: CharacterRepository) {
         return characterRepository.findAllCharactersByStatus(status)
     }
 
-    fun getCharacterById(characterId: Long): Optional<Character> {
+    fun getCharacterById(characterId: Long): Character {
         return characterRepository.findById(characterId)
+            .orElseThrow { EntityNotFoundException("Couldn't find character by id: $characterId") }
     }
 
     fun getAllCharacters(): List<Character> {
@@ -31,10 +33,26 @@ class CharacterService(private val characterRepository: CharacterRepository) {
     }
 
     fun updateCharacter(character: Character) {
-        characterRepository.save(character)
+        if (characterRepository.existsById(character.id)) {
+            characterRepository.save(Character(
+                character.id,
+                character.name,
+                character.status,
+                character.species,
+                character.origin,
+                character.image,
+                character.episodes
+            ))
+        } else {
+            throw EntityNotFoundException("Couldn't find character by id: ${character.id}")
+        }
     }
 
     fun deleteCharacter(characterId: Long) {
-        characterRepository.deleteById(characterId)
+        if (characterRepository.existsById(characterId)) {
+            characterRepository.deleteById(characterId)
+        } else {
+            throw EntityNotFoundException("Couldn't find character by id: $characterId")
+        }
     }
 }
