+++ Backend/src/main/kotlin/corp/siendev/com/rickandmorty/service/LocationService.kt
@@ -2,17 +2,19 @@ package corp.siendev.com.rickandmorty.service
 
 import corp.siendev.com.rickandmorty.entity.Location
 import corp.siendev.com.rickandmorty.repository.LocationRepository
+import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
-import java.util.Optional
 
 @Service
 class LocationService(private val locationRepository: LocationRepository) {
-    fun getLocationByName(locationName: String): Optional<Location> {
+    fun getLocationByName(locationName: String): Location {
         return locationRepository.findLocationByName(locationName)
+            .orElseThrow { EntityNotFoundException("Couldn't find location by location: $locationName") }
     }
 
-    fun getLocationById(id: Long): Optional<Location> {
+    fun getLocationById(id: Long): Location {
         return locationRepository.findById(id)
+            .orElseThrow { EntityNotFoundException("Couldn't find location by id: $id") }
     }
 
     fun getAllLocationsByDimension(dimension: String): List<Location> {
@@ -28,10 +30,24 @@ class LocationService(private val locationRepository: LocationRepository) {
     }
 
     fun updateLocation(location: Location) {
-        locationRepository.save(location)
+        if (locationRepository.existsById(location.id)) {
+            locationRepository.save(Location(
+                location.id,
+                location.name,
+                location.type,
+                location.dimension,
+                location.character
+            ))
+        } else {
+            throw EntityNotFoundException("Couldn't find location by id: ${location.id}")
+        }
     }
 
     fun deleteLocation(id: Long) {
-        locationRepository.deleteById(id)
+        if (locationRepository.existsById(id)) {
+            locationRepository.deleteById(id)
+        } else {
+            throw EntityNotFoundException("Couldn't find location by id: $id")
+        }
     }
 }
